+++ projects/raytracer/BRDFs/Lambertian.cpp
@@ -41,7 +41,7 @@ RGBColor Lambertian::f(const ShadeRec& sr, const Vector3d& wo, const Vector3d& w
 /* this generates a direction by sampling the hemisphere with a cosine distribution
    this is called in path_shade for any material with a diffuse shading component
    the samples have to be stored with a cosine distribution */
-RGBColor Lambertian::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi, float& pdf) const {
+RGBColor Lambertian::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi, double& pdf) const {
   Vector3d w = sr.normal;
   Vector3d v = Vector3d(0.0034, 1, 0.0071).cross(w);
   v.normalize();
