+++ projects/raytracer/BRDFs/PerfectSpecular.cpp
@@ -24,7 +24,7 @@ RGBColor PerfectSpecular::f(const ShadeRec& sr, const Vector3d& wo, const Vector
    it's called from from the functions Reflective::shade and Transparent::shade.
    the fabs in the last statement is for transparency */
 RGBColor PerfectSpecular::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi) const {
-  float ndotwo = sr.normal.dot(wo);
+  double ndotwo = sr.normal.dot(wo);
   wi = -wo + (2.0 * sr.normal * ndotwo);
   return kr * cr / fabs(sr.normal.dot(wi));
 }
@@ -32,8 +32,8 @@ RGBColor PerfectSpecular::sample_f(const ShadeRec& sr, const Vector3d& wo, Vecto
 
 /* this version of sample_f is used with path tracing
    it returns ndotwi in the pdf */
-RGBColor PerfectSpecular::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi, float& pdf) const {
-  float ndotwo = sr.normal.dot(wo);
+RGBColor PerfectSpecular::sample_f(const ShadeRec& sr, const Vector3d& wo, Vector3d& wi, double& pdf) const {
+  double ndotwo = sr.normal.dot(wo);
   wi = -wo + (2.0 * sr.normal * ndotwo);
   pdf = fabs(sr.normal.dot(wi));
   return kr * cr;
